// src/ui/client/client-form.types.ts
/* ------------------------------------------------------------------ */
/* Client form values                                                  */
/* ------------------------------------------------------------------ */

export interface ClientFormValues {
  nom: string;
  code: string;
  adresse: string;
  codePostal: string;
  ville: string;
  pays: string;
  email: string;
  telephone: string;
  actif: boolean;
}